# reducing number of maxHits in CA