--- a/RecoLocalTracker/SiPixelClusterizer/interface/PixelTrackingGPUConstants.h
+++ b/RecoLocalTracker/SiPixelClusterizer/interface/PixelTrackingGPUConstants.h
@@ -1,0 +1,10 @@
+#ifndef RecoLocalTracker_SiPixelClusterizer_PixelTrackingGPUConstants_h
+#define RecoLocalTracker_SiPixelClusterizer_PixelTrackingGPUConstants_h
+
+#include <cstdint>
+
+namespace GPUConstants {
+  constexpr uint16_t maxNumberOfHits = 20000;
+
+}
+#endif