--- a/RecoLocalTracker/SiPixelClusterizer/interface/PixelTrackingGPUConstants.h
+++ b/RecoLocalTracker/SiPixelClusterizer/interface/PixelTrackingGPUConstants.h
@@ -1,10 +1,11 @@
-#ifndef RecoLocalTracker_SiPixelClusterizer_PixelTrackingGPUConstants_h
-#define RecoLocalTracker_SiPixelClusterizer_PixelTrackingGPUConstants_h
+#ifndef RecoLocalTracker_SiPixelClusterizer_interface_PixelTrackingGPUConstants_h
+#define RecoLocalTracker_SiPixelClusterizer_interface_PixelTrackingGPUConstants_h
 
 #include <cstdint>
 
 namespace PixelGPUConstants {
-  constexpr uint16_t maxNumberOfHits = 30000;
+  constexpr uint16_t maxNumberOfHits = 40000;       // data at pileup 50 has 18300 +/- 3500 hits; 40000 is around 6 sigma away
 
 }
-#endif
+
+#endif // RecoLocalTracker_SiPixelClusterizer_interface_PixelTrackingGPUConstants_h