The maximum number of hits per layer for PU50 rarely exceeds 4500 in BPix1. 
The limit was set to 20k and now reduced to 6k.