@felicepantaleo 2018-09-11
:construction: Validation running at vinavx2.cern.ch:/data/user/fpantale/patatrack-validation/run.Be59wxIF1T ...