@fwyzard 2018-09-11
I agree that we should size the memory buffers to accommodate "most" of the events, rather than the possible theoretical maximum; however, we also need some way to handle the tails.