@felicepantaleo 2018-09-11
the approach now has changed. 
We have only one global array of hits and `isOuterHitOfCell`
So this is now increased to the maximum sum of the averages per layer instead of the sum of the maximum sizes per layer