@fwyzard 2018-09-12
The plots do show some small differences in the fakes and duplicates.
Can we assume them to be within the usual noise ?