@fwyzard 2018-09-13
@felicepantaleo is this in good shape to be merged, or do you expect further changes ?