@felicepantaleo 2018-09-13
You can merge, I will make a more generic optimization PR once I'm back to
CERN

On Thu, 13 Sep 2018, 14:51 Andrea Bocci, <notifications@github.com> wrote:

> @felicepantaleo <https://github.com/felicepantaleo> is this in good shape
> to be merged, or do you expect further changes ?
>
> —
> You are receiving this because you were mentioned.
> Reply to this email directly, view it on GitHub
> <https://github.com/cms-patatrack/cmssw/pull/163#issuecomment-420995268>,
> or mute the thread
> <https://github.com/notifications/unsubscribe-auth/AGSkg6DBG9TyhaIyEGz8Gs-A8nfB6DNwks5ualTcgaJpZM4WjGYO>
> .
>