@fwyzard 2018-09-13
Here's the distribution of the number of hits for the first ~2800 events:
![image](https://user-images.githubusercontent.com/4069793/45492642-18314300-b76d-11e8-98a6-30aab5ff85db.png)