@fwyzard 2018-09-13
The distribution for data is reasonably a Gaussian at 18347 ± 3480, so `30000` is at ~3.3 σ from the average.
To be on the safe side, I'd bump to `40000` which is ~6 σ away...