@fwyzard 2018-09-13
With the latest changes, I am able to run on a larger sample of data.
Comparing the performance with and without this PR, I see no changes in the kernel themselves:

**before**
```
cuobjdump -arch=sm_70 ...
 Function kernel_connect(GPU::SimpleVector<Quadruplet>*, GPUCACell*, unsigned int const*, GPU::VecArray<unsigned int, 256>*, float, float, float, float, float, unsigned int, unsigned int):
  REG:48 STACK:8 SHARED:0 LOCAL:0 CONSTANT[0]:412 TEXTURE:0 SURFACE:0 SAMPLER:0
 Function kernel_checkOverflows(GPU::SimpleVector<Quadruplet>*, GPUCACell*, unsigned int const*, GPU::VecArray<unsigned int, 256>*, unsigned int):
  REG:24 STACK:8 SHARED:0 LOCAL:0 CONSTANT[0]:388 TEXTURE:0 SURFACE:0 SAMPLER:0

nvprof:
                       Time     Calls       Avg       Min       Max  Name
                   132.65ms      1200  110.54us  25.311us  385.15us  kernel_connect(...)
                    21.77ms      1200   18.14us   5.280us   34.24us  kernel_checkOverflows(...)
```

**after**
```
cuobjdump -arch=sm_70 ...
 Function kernel_connect(GPU::SimpleVector<Quadruplet>*, GPUCACell*, unsigned int const*, GPU::VecArray<unsigned int, 256>*, float, float, float, float, float, unsigned int, unsigned int):
  REG:48 STACK:8 SHARED:0 LOCAL:0 CONSTANT[0]:412 TEXTURE:0 SURFACE:0 SAMPLER:0
 Function kernel_checkOverflows(GPU::SimpleVector<Quadruplet>*, GPUCACell*, unsigned int const*, GPU::VecArray<unsigned int, 256>*, unsigned int, unsigned int):
  REG:24 STACK:8 SHARED:0 LOCAL:0 CONSTANT[0]:392 TEXTURE:0 SURFACE:0 SAMPLER:0

nvprof:
                       Time     Calls       Avg       Min       Max  Name
                   132.01ms      1200  110.01us  26.912us  392.12us  kernel_connect(...)
                    21.32ms      1200   17.76us   5.184us   36.77us  kernel_checkOverflows(...)
```

However, I do observe a significant boost to the overall throughput:

**before**
  - mean:   551 ± 20 ev/s
  - best:   573 ± 22 ev/s

**after**
  - mean:   639 ± 20 ev/s
  - best:   641 ± 26 ev/s